+++ CHANGELOG.md
@@ -12,9 +12,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Converted SNP attestation UVM endorsements from integer to arbitrary string.
 
 - Add `/node/ready/app` and `/node/ready/gov` endpoints for the use of load balancers wanting to check if a node is ready to accept application or governance transactions. See [Operator RPC API](https://microsoft.github.io/CCF/main/operations/operator_rpc_api.html) for details.
-- Update `fmt` library from `9.1.0` to `10.1.1`.
+- Updated `llhttp` from `6.0.9` to `9.0.1`.
+- Updated `fmt` library from `9.1.0` to `10.1.1`.
 - Updated QCBOR from `1.1` to `1.2`.
-- Upgrade `nghttp2` from `1.51.0` to `1.55.1`.
+- Updated `nghttp2` from `1.51.0` to `1.55.1`.
 
 ## [4.0.7]